[FFM-92] +: Deleted scheduled_at at profile quotes, use scheduled_at at profile
The column is deleted, and we now use the scheduled_at date at the profile
for creating quotes and orders.
The date is calculated and set at the profile when the profile is created,
and after an order is created.
When creating quotes, the date at the profile is used when it's available,
else it will be calculated (and saved at the profile).
When calculating the scheduled date, the created_at date of the last order
of the profile will be used as the last schedule date, or if there's no order
yet, the created_at date of the profile.