[FFM-92] +: Reset quote profile addresses to order profile addresses when creating order
When a profile is edited, the profile addresses become of the type quote.
When creating a quote, we got the same addresses in the order, so we want
to save those addresses at the profile, because a quote can be deleted, and
with that the addresses, an order won't be deleted (in a normal scenario).