[FFM-99] B: Create profiles via cronjob instead of observer after invoice creation
The problem was that it can take up to a few seconds for Adyen to
create the billing agreement and give the correct responses, which
results in an error when creating the profile(s) that the billing
agreement couldn't be found. Now we wait untill the order is processing,
which means the billing agreement and the orders' invoice is created,
so we can create profile(s) without errors.