[FFM-145] B: Enterprise module XML file
This file is needed when running Enterprise, because the SalesArchive
module adds a union to the order collection of a customer, to add
the orders in the archive. We join the subscription ID in this
collection, and this must be done after the order archive is added
to the collection.