[FFM-197] B: Divide row total through original ordered qty
Follow up of d923d59acde23e2ebe15e8ada9e9e32a8761c27a;
When more than 1 of an item was ordered, the row total was a
multiplied value of the original product price. Now we divide
the row total that we save at the subscription item through
the ordered qty.
Example:
2x Product A (price 20)
Was saved at the subscription item as 40 (2x 20), so when a
quote was created of that subscription, the row total became
80 (2x 40). Now we divide 40 through 2 before saving at the
subscription, so a new quote will be 40 (2x 20).